requirements.txt
discord.py
discord.py[voice]
youtube_dl
PyNaCl
datetime
googletrans==3.1.0a0
twilio
python-decouple